refactor(langchain): address review feedback before merge
- Move detectProvider to langchain-utils.ts — keeps internal utility out
  of the public adapter API surface; runner and tests import from there
- Rename exportToLangChain -> exportToLangChainString — aligns with the
  String-suffix convention used by all other adapters
- Add fidelity-matrix.csv row for langchain adapter
