+++ src/adapters/index.ts
@@ -11,4 +11,4 @@ export { exportToGeminiString, exportToGemini } from './gemini.js';
 export { exportToCodexString, exportToCodex } from './codex.js';
 export { exportToKiroString, exportToKiro } from './kiro.js';
 export { exportToGitclawString, exportToGitclaw } from './gitclaw.js';
-export { exportToLangChain } from './langchain.js';
+export { exportToLangChainString } from './langchain.js';
